Move pagination to shared components
