+++ assets/stylesheets/_components.scss
@@ -198,6 +198,7 @@
 @import 'components/tinymce/style';
 @import 'components/title-format-editor/style';
 @import 'components/token-field/style';
+@import 'components/pagination/style';
 @import 'components/post-schedule/style';
 @import 'components/phone-input/style';
 @import 'components/section-header/style';
@@ -392,7 +393,6 @@
 @import 'my-sites/stats/all-time/style';
 @import 'my-sites/stats/geochart/style';
 @import 'my-sites/stats/most-popular/style';
-@import 'my-sites/stats/pagination/style';
 @import 'my-sites/stats/post-performance/style';
 @import 'my-sites/stats/post-trends/style';
 @import 'my-sites/stats/stats-chart-tabs/style';
