+++ client/components/pagination/style.scss
